+++ Application/Features/Categories/GetMainCategories/GetMainCategoriesHandler.cs
@@ -1,7 +1,7 @@
 using AutoMapper;
 using Domain.DTOs.Category;
 using Infrastructure.Interfaces.Repositories;
-using Infrastructure.Services.Caching;
+using Infrastructure.Interfaces.Services;
 using MediatR;
 
 namespace Application.Features.Categories.GetMainCategories;
